Changed GitHub repo language tag
This tag is automatically generated by GitHub and depends on relative file sizes in your repo. In our case, this flag was set to HTML because of the contents of the data subdir. We set this to R instead, by a command to ignore the contents of the data subdir when calculating relative file sizes.